+++ master/src/test/java/com/examples/with/different/packagename/testcarver/DateTimeConverter.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Copyright (C) 2010-2018 Gordon Fraser, Andrea Arcuri and EvoSuite
  * contributors
  *
@@ -230,7 +230,7 @@ public abstract class DateTimeConverter extends AbstractConverter {
         } else if (value instanceof Calendar) {
             date = ((Calendar)value).getTime();
         } else if (value instanceof Long) {
-            date = new Date(((Long)value).longValue());
+            date = new Date((Long) value);
         }
 
         String result = null;
@@ -311,7 +311,7 @@ public abstract class DateTimeConverter extends AbstractConverter {
         // Handle Long
         if (value instanceof Long) {
             Long longObj = (Long)value;
-            return toDate(targetType, longObj.longValue());
+            return toDate(targetType, longObj);
         }
 
         // Convert all other types to String & handle
